requirements.txt
pytorch
numpy
cv2
distance
nltk